+++ src/main/java/lu/itrust/business/TS/database/service/ServiceUserSqLite.java
@@ -15,11 +15,11 @@ import lu.itrust.business.TS.usermanagement.User;
  * @since Jan 16, 2013
  */
 public interface ServiceUserSqLite {
-	public UserSQLite get(Integer id);
+	public UserSQLite get(Long id);
 
 	public UserSQLite getByFilename(String filename);
 
-	public UserSQLite getByIdAndUser(Integer idFile, String username);
+	public UserSQLite getByIdAndUser(Long idFile, String username);
 
 	public List<UserSQLite> getAllFromUser(String username);
 
@@ -31,7 +31,7 @@ public interface ServiceUserSqLite {
 
 	public UserSQLite merge(UserSQLite userSqLite);
 
-	public void delete(Integer idUserSqLite);
+	public void delete(Long idUserSqLite);
 
 	public void delete(String filename);
